kinderbackend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
Pillow==12.0.0
pytest==7.4.3
httpx==0.25.2
python-multipart==0.0.6
psycopg[binary]==3.1.19
alembic==1.12.1
python-dotenv==1.0.0
stripe==14.0.1
redis==5.2.1
z-ai-web-dev-sdk
openai
